+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - (DEV) Set The `README` gif loops more times and waits longer between loops.
 - (DEV) Refactor set width/height in story.constructor.
 - Change from fixed to dynamic dimensions in `animate-badges/main.scss`.
+- Change `gallery/animate-badges` to support teleported dynamic story.
 
 ### Added
 - Add Google Analytics to Documentation.